clean up modularisation
moved "run_instantiation" from algorithms to implementation since it uses storage
